+++ app/auth/router.py
@@ -6,8 +6,8 @@ from fastapi import Depends, HTTPException, status, APIRouter
 from fastapi.security import OAuth2PasswordRequestForm
 from dotenv import load_dotenv
 
-from app.authentication import authenticate_user, create_access_token
-from app.models import Token
+from app.auth.services import authenticate_user, create_access_token
+from app.auth.models import Token
 
 load_dotenv()
 ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES"))
@@ -17,7 +17,7 @@ router = APIRouter(tags=["tokens"])
 
 @router.post("/login", response_model=Token)
 async def login_for_access_token(
-    form_data: Annotated[OAuth2PasswordRequestForm, Depends()]
+    form_data: Annotated[OAuth2PasswordRequestForm, Depends()],
 ):
     user = authenticate_user(form_data.username, form_data.password)
     if user is None:
@@ -36,7 +36,7 @@ async def login_for_access_token(
 
 @router.post("/register", response_model=Token)
 async def register_for_access_token(
-    form_data: Annotated[OAuth2PasswordRequestForm, Depends()]
+    form_data: Annotated[OAuth2PasswordRequestForm, Depends()],
 ):
     user = authenticate_user(form_data.username, form_data.password)
     if user is not None:
